requirements.txt
schedule==1.2.2
requests==2.32.5
python-dotenv==1.1.1
pyyaml==6.0.2
youtube-transcript-api==1.2.2
google-api-python-client==2.182.0
google-generativeai==0.8.5
yfinance==0.2.66
feedparser==6.0.12
colorlog==6.9.0
psutil==7.1.0
aiosqlite==0.20.0
